refactoring of class hierarchy
